chore: use swipe block hook added
